refactor: review cleanup — type safety, deduplication, and efficiency
- Replace all string literals with enum .as_str() calls (MessageStatus,
  MessageType, WorkStatus, RunStatus) for type safety
- Extract shared now_sql()/now_sql_nullable() helpers into db.rs
- Extract append_message() private helper in SessionStore (removes duplication)
- Extract dequeue_filtered() helper with bulk UPDATE via eq_any (fixes N+1)
- Use RETURNING clause in enqueue() (removes LastInsertRowId workaround)
- Collapse ensure_session + UPDATE into ON CONFLICT DO UPDATE upsert
- Make Database::with() pub(crate) to prevent abstraction leaks
- Align QueueMessage.id to i32 (matches diesel schema Integer type)

+++ crates/opengoose-teams/src/context.rs
@@ -62,7 +62,7 @@ impl OrchestrationContext {
     /// Convenience: read all broadcasts for this run since a given message ID.
     pub fn read_broadcasts(
         &self,
-        since_id: Option<i64>,
+        since_id: Option<i32>,
     ) -> Vec<opengoose_persistence::QueueMessage> {
         match self.queue().read_broadcasts(&self.team_run_id, since_id) {
             Ok(v) => v,
